+++ src/main/java/ua/shpp/dto/BranchResponseDTO.java
@@ -5,9 +5,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "Response branch")
 public record BranchResponseDTO(
         @Schema(description = "Id of branch", example = "1")
-        Integer id,
+        Long id,
         @Schema(description = "Id of organization", example = "1")
-        Integer organizationId,
+        Long organizationId,
         @Schema(description = "Name of branch")
         String name
 ) {
